Create new "wine" folder and move Wine Manager
